+++ test/mocks/video/message_box_test.cpp
@@ -33,11 +33,11 @@ TEST_F(MessageBoxTest, ShowStatic)
 
   {  // With defaults
     cen::message_box::show(title, message);
-    ASSERT_EQ(1, SDL_ShowSimpleMessageBox_fake.call_count);
+    ASSERT_EQ(1u, SDL_ShowSimpleMessageBox_fake.call_count);
 
 #if SDL_VERSION_ATLEAST(2, 0, 12)
     ASSERT_EQ(SDL_MESSAGEBOX_BUTTONS_LEFT_TO_RIGHT | SDL_MESSAGEBOX_INFORMATION,
-              SDL_ShowSimpleMessageBox_fake.arg0_val);
+              static_cast<int>(SDL_ShowSimpleMessageBox_fake.arg0_val));
 #endif  // SDL_VERSION_ATLEAST(2, 0, 12)
 
     ASSERT_STREQ(title.c_str(), SDL_ShowSimpleMessageBox_fake.arg1_val);
@@ -47,11 +47,11 @@ TEST_F(MessageBoxTest, ShowStatic)
 
   {  // With custom type
     cen::message_box::show(title, message, cen::message_box_type::error);
-    ASSERT_EQ(2, SDL_ShowSimpleMessageBox_fake.call_count);
+    ASSERT_EQ(2u, SDL_ShowSimpleMessageBox_fake.call_count);
 
 #if SDL_VERSION_ATLEAST(2, 0, 12)
     ASSERT_EQ(SDL_MESSAGEBOX_BUTTONS_LEFT_TO_RIGHT | SDL_MESSAGEBOX_ERROR,
-              SDL_ShowSimpleMessageBox_fake.arg0_val);
+              static_cast<int>(SDL_ShowSimpleMessageBox_fake.arg0_val));
 #endif  // SDL_VERSION_ATLEAST(2, 0, 12)
 
     ASSERT_STREQ(title.c_str(), SDL_ShowSimpleMessageBox_fake.arg1_val);
@@ -64,11 +64,11 @@ TEST_F(MessageBoxTest, ShowStatic)
                            message,
                            cen::message_box_type::error,
                            cen::button_order::right_to_left);
-    ASSERT_EQ(3, SDL_ShowSimpleMessageBox_fake.call_count);
+    ASSERT_EQ(3u, SDL_ShowSimpleMessageBox_fake.call_count);
 
 #if SDL_VERSION_ATLEAST(2, 0, 12)
     ASSERT_EQ(SDL_MESSAGEBOX_BUTTONS_RIGHT_TO_LEFT | SDL_MESSAGEBOX_ERROR,
-              SDL_ShowSimpleMessageBox_fake.arg0_val);
+              static_cast<int>(SDL_ShowSimpleMessageBox_fake.arg0_val));
 #endif  // SDL_VERSION_ATLEAST(2, 0, 12)
 
     ASSERT_STREQ(title.c_str(), SDL_ShowSimpleMessageBox_fake.arg1_val);
@@ -82,9 +82,9 @@ TEST_F(MessageBoxTest, Show)
   cen::message_box mb;
 
   mb.show();
-  ASSERT_EQ(1, SDL_ShowMessageBox_fake.call_count);
+  ASSERT_EQ(1u, SDL_ShowMessageBox_fake.call_count);
 
   cen::window_handle window{nullptr};
   mb.show(window);
-  ASSERT_EQ(2, SDL_ShowMessageBox_fake.call_count);
+  ASSERT_EQ(2u, SDL_ShowMessageBox_fake.call_count);
 }
